+++ src/components/CartWidget.jsx
@@ -1,0 +1,13 @@
+  const CartWidget = () => {
+    return(
+  <div className="d-flex">
+      <button type="button" className="btn btn-primary position-relative align-items-center justify-content-center logo2"><img className="m-2 logo" src="/image/logos/logsteam.png" alt="Logo" /> Inciar Sesion<span className="position-absolute top-0 start-100 translate-middle badge rounded-pill bg-danger">1</span></button>
+
+      
+  </div>    
+      
+      )
+
+  };
+
+export default CartWidget;
